function moved to utils
