+++ src/components/Header/Header.tsx
@@ -1,30 +1,21 @@
 import * as React from 'react'
 import { NavMenu } from 'src/components/NavMenu/NavMenu'
 import { TextDisplay } from 'src/components/TextDisplay/TextDisplay'
-import { SearchBar } from './SearchBar'
-import Cross from 'src/assets/cross.svg'
-import Search from 'src/assets/search.svg'
 import Menu from 'src/assets/menu.svg'
 import styles from './Header.module.scss'
 import { SignUpLoginLayout } from '../Modal/SignUpLoginModal'
 import { UserContext } from 'src/context/UserContext'
 import { Link } from 'gatsby'
+import { fadeIn } from '../../utils/Modal'
 
 export const Header = () => {
-  const { user } = React.useContext(UserContext)
-  const [showSearchBar, setShowSearchBar] = React.useState(false)
+  const [] = React.useState(false)
   const [showNavMenu, setShowNavMenu] = React.useState(false)
   const [showLoginModal, setShowLoginModal] = React.useState(false)
   const [showSignUpModal, setShowSignUpModal] = React.useState(false)
 
   React.useEffect(() => {
-    setTimeout(() => {
-      const element = document.getElementsByClassName('modal-transition')[0]
-      if (element) {
-        element.classList.remove('close')
-        element.classList.add('open')
-      }
-    }, 100)
+    fadeIn()
   }, [showLoginModal])
 
   return (
