+++ src/components/Modal/LoginModal.tsx
@@ -9,6 +9,7 @@ import { CompleteModal } from './CompleteModal'
 import { ErrorModal } from './ErrorModal'
 import { UserContext } from 'src/context/UserContext'
 import { AuthService } from 'src/services/AuthService'
+import { fadeIn } from '../../utils/Modal'
 
 type LoginComponentProps = ModalProps & {
   showSignUpModal: VoidFunction
@@ -20,13 +21,7 @@ const LoginComponent: React.FC<LoginComponentProps> = ({ closeModal, showSignUpM
   const [errorMesasge, setErrorMessage] = React.useState('')
 
   React.useEffect(() => {
-    setTimeout(() => {
-      const element = document.getElementsByClassName('modal-transition')[1]
-      if (element) {
-        element.classList.remove('close')
-        element.classList.add('open')
-      }
-    }, 100)
+    fadeIn();
   }, [user, errorMesasge])
 
   return (
